1 - Modularize agents
